Add codedark repo to vim plugins list
Vim airline plugin was no longer auto-magically doing the colorscheme
based off of the vim colorscheme, so needed to pull down the repo to
include the airline theme.

May not keep it because I have customizations to the codedark
colorscheme (just two minor ones) ... and it can be confusing why I have
it in two places.